wrapper script: do not use tmp dir for the .so
Launcher directory cannot be used because it contains space, which
cannot be included in LD_PRELOAD. Temporary directory doesn't work
because Yousician launcher exits while Yousician starts and so does the
script, cleaning the temporary files.

Instead of linking yousician_transcode.so to $TMPDIR, put it into
~/.config/unity3d/Yousician/Yousician – this should be safe and it won't
litter the filesystem.